Restructure files to align with vertical slice architecture: - Removed layout-specific sub-directories like 'right-side-bar' and 'left-side-bar'. - Moved 'hot-ideas.ts', 'related-ideas.ts', and 'top-supporters.ts' from 'right-side-bar' to 'src/features/idea/components'. - Moved 'create-idea.ts', 'create-solution.ts', 'discover-page.ts', 'edit-profile.ts', 'home-page.ts', 'idea-page.ts', and 'view-profile.ts' from 'src/pages' to 'src/features/pages'. - Moved 'my-ideas.ts' and 'my-solutions.ts' from 'left-side-bar' to 'src/features/user/components'.
