+++ src/features/idea/components/hot-ideas.ts
