+++ src/features/layout/components/right-side-bar.ts
@@ -6,9 +6,9 @@
 import { LitElement, html, css } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 
-import '@/features/idea/components/right-side-bar/hot-ideas';
-import '@/features/idea/components/right-side-bar/related-ideas';
-import '@/features/idea/components/right-side-bar/top-supporters';
+import '@/features/idea/components/hot-ideas';
+import '@/features/idea/components/related-ideas';
+import '@/features/idea/components/top-supporters';
 import '@/features/tags/components/watched-tags';
 import '@/features/tags/components/popular-tags';
